# `wp_theme_has_theme_json`: add public method to know whether the theme supports `theme.json`