--- a/src/wp-includes/class-wp-theme-json-resolver.php
+++ b/src/wp-includes/class-wp-theme-json-resolver.php
@@ -58,14 +58,6 @@ class WP_Theme_JSON_Resolver {
 	protected static $theme = null;
 
 	/**
-	 * Whether or not the theme supports theme.json.
-	 *
-	 * @since 5.8.0
-	 * @var bool
-	 */
-	protected static $theme_has_support = null;
-
-	/**
 	 * Container for data coming from the user.
 	 *
 	 * @since 5.9.0
@@ -295,7 +287,7 @@ class WP_Theme_JSON_Resolver {
 		 * and merge the static::$theme upon that.
 		 */
 		$theme_support_data = WP_Theme_JSON::get_from_editor_settings( get_default_block_editor_settings() );
-		if ( ! static::theme_has_support() ) {
+		if ( ! wp_theme_has_theme_json() ) {
 			if ( ! isset( $theme_support_data['settings']['color'] ) ) {
 				$theme_support_data['settings']['color'] = array();
 			}
@@ -416,11 +408,11 @@ class WP_Theme_JSON_Resolver {
 		/*
 		 * Bail early if the theme does not support a theme.json.
 		 *
-		 * Since WP_Theme_JSON_Resolver::theme_has_support() only supports the active
+		 * Since wp_theme_has_theme_json() only supports the active
 		 * theme, the extra condition for whether $theme is the active theme is
 		 * present here.
 		 */
-		if ( $theme->get_stylesheet() === get_stylesheet() && ! static::theme_has_support() ) {
+		if ( $theme->get_stylesheet() === get_stylesheet() && ! wp_theme_has_theme_json() ) {
 			return array();
 		}
 
@@ -597,18 +589,14 @@ class WP_Theme_JSON_Resolver {
 	 *
 	 * @since 5.8.0
 	 * @since 5.9.0 Added a check in the parent theme.
+	 * @deprecated 6.2.0 Use wp_theme_has_theme_json() instead.
 	 *
 	 * @return bool
 	 */
 	public static function theme_has_support() {
-		if ( ! isset( static::$theme_has_support ) ) {
-			static::$theme_has_support = (
-				static::get_file_path_from_theme( 'theme.json' ) !== '' ||
-				static::get_file_path_from_theme( 'theme.json', true ) !== ''
-			);
-		}
+		_deprecated_function( __METHOD__, '6.2.0', 'wp_theme_has_theme_json()' );
 
-		return static::$theme_has_support;
+		return wp_theme_has_theme_json();
 	}
 
 	/**
@@ -651,7 +639,6 @@ class WP_Theme_JSON_Resolver {
 		static::$theme                    = null;
 		static::$user                     = null;
 		static::$user_custom_post_type_id = null;
-		static::$theme_has_support        = null;
 		static::$i18n_schema              = null;
 	}