--- a/tests/phpunit/data/themedir1/default-child-no-theme-json/style.css
+++ b/tests/phpunit/data/themedir1/default-child-no-theme-json/style.css
@@ -1,0 +1,8 @@
+/*
+Theme Name: Default Child Theme with no theme.json
+Theme URI: https://wordpress.org/
+Description: For testing purposes only.
+Template: default
+Version: 1.0.0
+Text Domain: default-child-no-theme-json
+*/