@oandregal 2022-11-16
https://github.com/WordPress/gutenberg/pull/45543 is updating a bit the code, I need to port those changes here once they land in Gutenberg.