@spacedmonkey 2022-11-16
I think we should look at and focus on my PR here - https://github.com/WordPress/wordpress-develop/pull/3604