@oandregal 2022-12-13
I've ported the last changes done in Gutenberg to this PR, so it's up to date and ready to review. Marked the old conversations as done as they no longer reflected the state of this code (nor the conversations over at Gutenberg).

Because its relationship with [this other PR](https://github.com/WordPress/wordpress-develop/pull/3712) (both use the `theme_json` cache group: they need to make it persistent and offer a way to clean the data), I've ported the same changes over here at https://github.com/WordPress/wordpress-develop/pull/3556/commits/073766ed9300eb5be7b752c1c23a3b411173aa35 Whatever PR lands first, some changes will not be longer need it. I'm documenting them as PR comments.

cc @felixarntz @spacedmonkey @hellofromtonya @azaozz @anton-vlasenko @costdev