@oandregal 2022-12-14
@costdev Thanks for the suggestion to use a data provider in the tests, they're much clearer now! All your feedback was addressed.