@oandregal 2022-12-16
Rebased from latest `trunk` to fix some conflicts. I guess this is ready for merging if a core committer finds the time :)