@mcsf 2022-12-19
> Blocking this commit until we resolve [WordPress/gutenberg#45955](https://github.com/WordPress/gutenberg/pull/45955).
> 
> This PR changes the cache invalidation.

One PR is a backport of a function introduced and validated in Gutenberg. The other is a potential improvement still in debate. Both rely on `WP_Object_Cache`, differing only in the details. I don't see how one should block the other.