@hellofromtonya 2022-12-19
I'm confused why this backport PR includes caching changes as these seem out-of-scope for this initiative. Adding `wp_theme_has_theme_json()` and applying its changes are a specific scope of work for a single commit.

Does adding `wp_theme_has_theme_json()` not work without the caching changes? Or are these truly separate scopes of work?