@oandregal 2022-12-21
> I'm confused why this backport PR includes caching changes as these seem out-of-scope for this initiative. Adding wp_theme_has_theme_json() and applying its changes are a specific scope of work for a single commit.
> 
> Does adding wp_theme_has_theme_json() not work without the caching changes? Or are these truly separate scopes of work?

That's a great question. The more important thing for me is to have this method available. Not making performance worse is also something I tried to keep in mind, that's why I thought it should have _some_ cache, as the older one also had.

However! I also want to provide a more nuanced answer based on data. I ran some numbers and this is what I found (see [raw data](https://docs.google.com/spreadsheets/d/18f-m2KfPml6Bceo_4ehEwJ-NzHKW8rvo0FGhEzTf_I4/edit#gid=922364294) extracted from the cachegrind files linked below):

Approach | Cost in the full request (ms) | Cachegrind files
--- | --- | ---
Current status: `WP_Theme_JSON_Resolver::has_theme_support` | 0.05ms | [cachegrind.out.hello-world.trunk.zip](https://github.com/WordPress/wordpress-develop/files/10279587/cachegrind.out.hello-world.trunk.zip)
This PR: `wp_theme_has_theme_json` with object cache | 2.7ms | [cachegrind.out.hello-world.pr.zip](https://github.com/WordPress/wordpress-develop/files/10279593/cachegrind.out.hello-world.pr.zip)
`wp_theme_has_theme_json` using static variables as cache | 0.05ms | [cachegrind.out.hello-world.cachestatic.zip](https://github.com/WordPress/wordpress-develop/files/10279590/cachegrind.out.hello-world.cachestatic.zip)
`wp_theme_has_theme_json` with no cache (no object cache or static variables) | 1.13ms | [cachegrind.out.hello-world.nocache.zip](https://github.com/WordPress/wordpress-develop/files/10279592/cachegrind.out.hello-world.nocache.zip):

When looking at the numbers, you can say that using static variables is faster, and the `wp_theme_has_theme_json` without any cache performs faster than using the object cache :upside_down_face:

However, that's a partial picture. This is the other part of the picture:

- Not using any cache is vulnerable to grow conditions (the method can be used in more places, hence the performance impact will have a linear growth). Besides, the data I've shared only shows the performance _in my machine_ (some servers may have slower file access, different PHP versions, etc.).
- Using static variables in the method for caching comes with downsides in terms of API ergonomics: the cache leaks through the method parameters. In other words, you need to have this: `wp_theme_has_theme_json( $cache_clear )`. I actually started with this approach ([see thread](https://github.com/WordPress/wordpress-develop/pull/3556#discussion_r1012686381)).
- Using the object cache has the advantage that you can swap it for a faster implementation (plugins, etc.). Also, in the future, if/when we make the `theme_json` group persistent across requests, the cost is minimized (~1ms less at a minimum given if would not have to recalculate support).

All things considered, even the fact that we're talking 1ms here while there are PRs such as https://github.com/WordPress/wordpress-develop/pull/3789 that reduce +350ms, I think the current approach is the more future-proof.

With what I've learned, I wouldn't mind removing the cache altogether if that is preferable for core folks and provided it helps to ship this faster: it is no big impact either way. I wouldn't recommend going back to caching using static variables in the method, as it compromises the API.