@spacedmonkey 2023-01-16
Everyone reviewing this PR should consider this alternative https://github.com/WordPress/wordpress-develop/pull/3604. 

This has number of key benefits. 
- You can see if ANY theme support theme json. 
- Reuses the theme cache group. 
- Cleaner cache invalidation.